+++ gpkit/tests/t_constraints.py
@@ -34,5 +34,5 @@ class TestMonoEQConstraint(unittest.TestCase):
 TESTS = [TestConstraint, TestMonoEQConstraint]
 
 if __name__ == '__main__':
-    from gpkit.tests.run_tests import run_tests
+    from gpkit.tests.helpers import run_tests
     run_tests(TESTS)
