+++ gpkit/tests/t_geometric_program.py
@@ -234,5 +234,5 @@ for testcase in TEST_CASES:
             TESTS.append(test)
 
 if __name__ == "__main__":
-    from gpkit.tests.run_tests import run_tests
+    from gpkit.tests.helpers import run_tests
     run_tests(TESTS)
