+++ gpkit/tests/t_gp_solution_array.py
@@ -74,5 +74,5 @@ class TestGPSolutionArray(unittest.TestCase):
 TESTS = [TestGPSolutionArray]
 
 if __name__ == '__main__':
-    from gpkit.tests.run_tests import run_tests
+    from gpkit.tests.helpers import run_tests
     run_tests(TESTS)
