+++ gpkit/tests/t_posy_array.py
@@ -85,5 +85,5 @@ class TestPosyArray(unittest.TestCase):
 TESTS = [TestPosyArray]
 
 if __name__ == '__main__':
-    from gpkit.tests.run_tests import run_tests
+    from gpkit.tests.helpers import run_tests
     run_tests(TESTS)
