+++ gpkit/tests/t_sub.py
@@ -207,5 +207,5 @@ class TestGPSubs(unittest.TestCase):
 TESTS = [TestNomialSubs, TestGPSubs]
 
 if __name__ == '__main__':
-    from gpkit.tests.run_tests import run_tests
+    from gpkit.tests.helpers import run_tests
     run_tests(TESTS)
