+++ gpkit/tests/t_vars.py
@@ -116,5 +116,5 @@ class T_VectorVariable(unittest.TestCase):
 TESTS = [T_VarKey, T_Variable, T_VectorVariable]
 
 if __name__ == '__main__':
-    from gpkit.tests.run_tests import run_tests
+    from gpkit.tests.helpers import run_tests
     run_tests(TESTS)
